feat(discord): add discord logger error service
