+++ src/enums/service-name.enum.spec.ts
@@ -163,6 +163,14 @@ describe(`ServiceNameEnum`, (): void => {
     );
   });
 
+  it(`should have a member "DISCORD_LOGGER_ERROR_SERVICE"`, (): void => {
+    expect.assertions(1);
+
+    expect(ServiceNameEnum.DISCORD_LOGGER_ERROR_SERVICE).toStrictEqual(
+      `DiscordLoggerErrorService`
+    );
+  });
+
   it(`should have a member "DISCORD_MENTION_SERVICE"`, (): void => {
     expect.assertions(1);
