+++ src/enums/service-name.enum.ts
@@ -20,6 +20,7 @@ export enum ServiceNameEnum {
   DISCORD_GUILD_MEMBER_ADD_SERVICE = `DiscordGuildMemberAddService`,
   DISCORD_GUILD_SERVICE = `DiscordGuildService`,
   DISCORD_LOGGER_SERVICE = `DiscordLoggerService`,
+  DISCORD_LOGGER_ERROR_SERVICE = `DiscordLoggerErrorService`,
   DISCORD_MENTION_SERVICE = `DiscordMentionService`,
   DISCORD_MESSAGE_AUTHOR_SERVICE = `DiscordMessageAuthorService`,
   DISCORD_MESSAGE_COMMAND_COOKIE_SERVICE = `DiscordMessageCommandCookieService`,
