+++ src/features/discord/logger/services/discord-logger.service.ts
@@ -1,25 +1,12 @@
-import {
-  Client,
-  EmbedFieldData,
-  MessageEmbedAuthor,
-  MessageEmbedFooter,
-  MessageEmbedOptions,
-  MessageEmbedThumbnail,
-} from "discord.js";
+import { Client } from "discord.js";
 import _ from "lodash";
-import moment from "moment";
 import { AbstractService } from "../../../../classes/abstract.service";
 import { ServiceNameEnum } from "../../../../enums/service-name.enum";
-import { ellipsis } from "../../../../functions/formatters/ellipsis";
 import { wrapInQuotes } from "../../../../functions/formatters/wrap-in-quotes";
 import { ChalkService } from "../../../logger/services/chalk.service";
 import { LoggerService } from "../../../logger/services/logger.service";
-import { DiscordGuildSoniaChannelNameEnum } from "../../guilds/enums/discord-guild-sonia-channel-name.enum";
-import { DiscordGuildSoniaService } from "../../guilds/services/discord-guild-sonia.service";
-import { IDiscordMessageResponse } from "../../messages/interfaces/discord-message-response";
-import { DiscordMessageConfigService } from "../../messages/services/config/discord-message-config.service";
 import { DiscordClientService } from "../../services/discord-client.service";
-import { DiscordSoniaService } from "../../users/services/discord-sonia.service";
+import { DiscordLoggerErrorService } from "./discord-logger-error.service";
 
 export class DiscordLoggerService extends AbstractService {
   private static _instance: DiscordLoggerService;
@@ -35,9 +22,7 @@ export class DiscordLoggerService extends AbstractService {
   public readonly discordClient: Client = DiscordClientService.getInstance().getClient();
   private readonly _loggerService: LoggerService = LoggerService.getInstance();
   private readonly _chalkService: ChalkService = ChalkService.getInstance();
-  private readonly _discordGuildSoniaService: DiscordGuildSoniaService = DiscordGuildSoniaService.getInstance();
-  private readonly _discordSoniaService: DiscordSoniaService = DiscordSoniaService.getInstance();
-  private readonly _discordMessageConfigService: DiscordMessageConfigService = DiscordMessageConfigService.getInstance();
+  private readonly _discordLoggerErrorService: DiscordLoggerErrorService = DiscordLoggerErrorService.getInstance();
 
   protected constructor() {
     super(ServiceNameEnum.DISCORD_LOGGER_SERVICE);
@@ -69,109 +54,12 @@ export class DiscordLoggerService extends AbstractService {
 
   private _listenForErrors(): void {
     this.discordClient.on(`error`, (error: Readonly<Error>): void => {
-      this._handleError(error);
+      this._discordLoggerErrorService.handleError(error);
     });
 
     this._loggerService.debug({
       context: this._serviceName,
       message: this._chalkService.text(`listen ${wrapInQuotes(`error`)} event`),
     });
-  }
-
-  private _handleError(error: Readonly<Error>): void {
-    this._loggerService.error({
-      context: this._serviceName,
-      message: this._chalkService.text(error),
-    });
-    this._loggerService.debug({
-      context: this._serviceName,
-      message: this._chalkService.text(
-        `send message to Sonia Discord errors channel`
-      ),
-    });
-
-    this._discordGuildSoniaService.sendMessageToChannel({
-      channelName: DiscordGuildSoniaChannelNameEnum.ERRORS,
-      messageResponse: this._getErrorMessageResponse(error),
-    });
-  }
-
-  private _getErrorMessageResponse(
-    error: Readonly<Error>
-  ): IDiscordMessageResponse {
-    return {
-      options: {
-        embed: this._getMessageEmbed(error),
-        split: true,
-      },
-      response: ``,
-    };
-  }
-
-  private _getMessageEmbed(error: Readonly<Error>): MessageEmbedOptions {
-    return {
-      author: this._getMessageEmbedAuthor(),
-      color: this._getMessageEmbedColor(),
-      description: this._getMessageEmbedDescription(error),
-      fields: this._getMessageEmbedFields(error),
-      footer: this._getMessageEmbedFooter(),
-      thumbnail: this._getMessageEmbedThumbnail(),
-      timestamp: this._getMessageEmbedTimestamp(),
-      title: this._getMessageEmbedTitle(error),
-    };
-  }
-
-  private _getMessageEmbedAuthor(): MessageEmbedAuthor {
-    return this._discordSoniaService.getCorporationMessageEmbedAuthor();
-  }
-
-  private _getMessageEmbedThumbnail(): MessageEmbedThumbnail {
-    return {
-      url: this._discordMessageConfigService.getMessageCommandErrorImageUrl(),
-    };
-  }
-
-  private _getMessageEmbedFooter(): MessageEmbedFooter {
-    const soniaImageUrl:
-      | string
-      | null = this._discordSoniaService.getImageUrl();
-
-    return {
-      iconURL: soniaImageUrl || undefined,
-      text: `Discord error`,
-    };
-  }
-
-  private _getMessageEmbedColor(): number {
-    return this._discordMessageConfigService.getMessageCommandErrorImageColor();
-  }
-
-  private _getMessageEmbedTimestamp(): Date {
-    return moment().toDate();
-  }
-
-  private _getMessageEmbedTitle(error: Readonly<Error>): string {
-    return error.name;
-  }
-
-  private _getMessageEmbedDescription(error: Readonly<Error>): string {
-    return error.message;
-  }
-
-  private _getMessageEmbedFields(error: Readonly<Error>): EmbedFieldData[] {
-    const embedFieldData: EmbedFieldData[] = [];
-
-    if (!_.isNil(error.stack)) {
-      embedFieldData.push(this._getMessageEmbedFieldError(error.stack));
-    }
-
-    return embedFieldData;
-  }
-
-  private _getMessageEmbedFieldError(stack: Readonly<string>): EmbedFieldData {
-    return {
-      name: `My blood trace`,
-      value: ellipsis(stack),
-    };
   }
 }
